fix - Cycles and some refactoring
