+++ src/screens/WalletScreen/AppTour/Context/AppTourContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const AppTourContext = createContext<any>(null);
